port serial code to boost asio, fix some packet stuff
git-svn-id: https://localhost/igvc@700 b8439097-6f30-0410-869d-b57f63bf1bb9

+++ 2010/software/trunk/arduino/standard_types.hpp
@@ -14,8 +14,8 @@
 	typedef char int8;
 	typedef unsigned char uint8;
 
-	typedef short int16;
-	typedef unsigned short uint16;
+	typedef int int16;
+	typedef unsigned int uint16;
 
 	typedef long int32;
 	typedef unsigned long uint32;
